Add files via upload
You can access the project files, including the data used for analysis, through the following Google Drive link:

https://shorturl.at/KZdJa

Please note that this link will take you to a Google Sheet containing the dataset used for the Excel-based Road Accident Dashboard. You can download the dataset and use it to replicate the steps described in the project.